+++ grails-app/services/de/iteratec/osm/measurement/environment/wptserver/WptInstructionService.groovy
@@ -17,6 +17,7 @@
 
 package de.iteratec.osm.measurement.environment.wptserver
 
+import de.iteratec.osm.ConfigService
 import de.iteratec.osm.measurement.environment.Location
 import de.iteratec.osm.measurement.environment.WebPageTestServer
 import de.iteratec.osm.measurement.schedule.Job
@@ -25,11 +26,25 @@ import de.iteratec.osm.result.WptStatus
 import de.iteratec.osm.util.PerformanceLoggingService
 import grails.async.Promise
 import groovy.util.slurpersupport.GPathResult
+import groovy.util.slurpersupport.NodeChild
 import groovyx.net.http.NativeHandlers
 
 import java.util.concurrent.locks.ReentrantLock
 
+import static de.iteratec.osm.util.PerformanceLoggingService.LogLevel.DEBUG
 import static grails.async.Promises.task
+
+class JobExecutionException extends Exception {
+    WptStatus wptStatus
+    String testId
+
+    JobExecutionException(String message, WptStatus wptStatus, String testId) {
+        super(message)
+        this.wptStatus = wptStatus
+        this.testId = testId
+    }
+
+}
 
 interface iResultListener {
     public String getListenerName()
@@ -67,6 +82,7 @@ class WptInstructionService {
 
     HttpRequestService httpRequestService
     PerformanceLoggingService performanceLoggingService
+    ConfigService configService
 
     /**
      * Listeners can register as oberservers.
@@ -88,15 +104,37 @@ class WptInstructionService {
      * 			Should contain all necessary for running tests on wptserver.
      * @return
      */
-    Object runtest(WebPageTestServer wptserver, Map params) {
-        Map paramsNotNull = params.findAll {k,v-> v != null  }
-        return httpRequestService.getRestClientFrom(wptserver).post {
-            request.uri.path = '/runtest.php'
-            request.body = paramsNotNull
-            request.contentType = 'application/x-www-form-urlencoded'
-            request.encoder('application/x-www-form-urlencoded', NativeHandlers.Encoders.&form)
-            request.headers['Accept'] = 'application/xml'
+    String runtest(Job job, int priority) {
+        WebPageTestServer wptServer = job?.location?.wptServer
+        if (!wptServer) {
+            throw new JobExecutionException("Missing wptServer in job ${job.label}", WptStatus.UNKNOWN, null)
         }
+
+        def parameters = fillRequestParameters(job)
+        parameters['priority'] = priority
+        Map paramsNotNull = parameters.findAll { k, v -> v != null }
+        def result = null
+        performanceLoggingService.logExecutionTime(DEBUG, "Launching job ${job.label}: Calling initial runtest on wptserver.", 1) {
+            result = httpRequestService.getRestClientFrom(wptServer).post {
+                request.uri.path = '/runtest.php'
+                request.body = paramsNotNull
+                request.contentType = 'application/x-www-form-urlencoded'
+                request.encoder('application/x-www-form-urlencoded', NativeHandlers.Encoders.&form)
+                request.headers['Accept'] = 'application/xml'
+            }
+        }
+
+        NodeChild runtestResponseXml = parseResponse(result, wptServer)
+        WptStatus wptStatus = WptStatus.byResultCode(runtestResponseXml?.statusCode?.toInteger())
+        String testId = runtestResponseXml?.data?.testId
+        if (wptStatus != WptStatus.COMPLETED) {
+            throw new JobExecutionException("Got status code ${wptStatus} from wptserver ${wptServer}", wptStatus, testId)
+        }
+        if (!testId) {
+            throw new JobExecutionException("Jobrun failed for: wptserver=${wptServer}, sent params=${parameters} => got no testId in response", wptStatus, testId)
+        }
+        log.info("Jobrun successfully launched: wptserver=${wptServer}, sent params=${parameters}, got testID: ${testId}")
+        return testId
     }
 
     Object cancelTest(WebPageTestServer wptserver, Map params) {
@@ -209,4 +247,119 @@ class WptInstructionService {
         )
     }
 
+    /**
+     * Maps the properties of a Job to the parameters expected by
+     * the REST API available at https://sites.google.com/a/webpagetest.org/docs/advanced-features/webpagetest-restful-apis
+     *
+     * @return A map of parameters suitable for POSTing a valid request to runtest.php
+     */
+    private Map fillRequestParameters(Job job) {
+        Map parameters = [
+                url            : '',
+                label          : job.label,
+                location       : job.location.uniqueIdentifierForServer,
+                runs           : job.runs,
+                fvonly         : job.firstViewOnly,
+                video          : job.captureVideo,
+                web10          : job.web10,
+                noscript       : job.noscript,
+                clearcerts     : job.clearcerts,
+                ignoreSSL      : job.ignoreSSL,
+                standards      : job.standards,
+                tcpdump        : job.tcpdump,
+                continuousVideo: job.continuousVideo,
+                private        : job.isPrivate,
+                block          : job.urlsToBlock,
+                mobile         : job.emulateMobile,
+                dpr            : job.devicePixelRation,
+                cmdline        : job.cmdlineOptions,
+                custom         : job.customMetrics,
+                tester         : job.tester,
+                timeline       : job.captureTimeline,
+                timelineStack  : job.javascriptCallstack,
+                mobileDevice   : job.mobileDevice,
+                lighthouse     : job.performLighthouseTest,
+                type           : job.optionalTestTypes,
+                customHeaders  : job.customHeaders,
+                trace          : job.trace,
+                traceCategories: job.traceCategories,
+                spof           : job.spof
+        ]
+        if (job.takeScreenshots == Job.TakeScreenshots.NONE) {
+            parameters.noimages = true
+        } else if (job.takeScreenshots == Job.TakeScreenshots.FULL) {
+            parameters.pngss = true
+        } else {
+            parameters.iq = job.imageQuality
+        }
+
+        if (job.saveBodies == Job.SaveBodies.HTML) {
+            parameters.htmlbody = true
+        } else if (job.saveBodies == Job.SaveBodies.ALL) {
+            parameters.bodies = true
+        }
+
+        if (configService.globalUserAgentSuffix && job.useGlobalUASuffix) {
+            parameters.appendua = configService.globalUserAgentSuffix
+        } else {
+            if (job.userAgent == Job.UserAgent.ORIGINAL) {
+                parameters.keepua = true
+            } else if (job.userAgent == Job.UserAgent.APPEND) {
+                parameters.appendua = job.appendUserAgent
+            } else if (job.userAgent == Job.UserAgent.OVERWRITE) {
+                parameters.uastring = job.userAgentString
+            }
+        }
+
+
+        if (job.noTrafficShapingAtAll) {
+            parameters.location += ".Native"
+        } else {
+            parameters.location += ".custom"
+            if (job.customConnectivityProfile) {
+                parameters.bwDown = job.bandwidthDown
+                parameters.bwUp = job.bandwidthUp
+                parameters.latency = job.latency
+                parameters.plr = job.packetLoss
+            } else {
+                parameters.bwDown = job.connectivityProfile.bandwidthDown
+                parameters.bwUp = job.connectivityProfile.bandwidthUp
+                parameters.latency = job.connectivityProfile.latency
+                parameters.plr = job.connectivityProfile.packetLoss
+            }
+        }
+
+        if (job.script) {
+            parameters.script = job.script.getParsedNavigationScript(job)
+            if (job.provideAuthenticateInformation) {
+                parameters.login = job.authUsername
+                parameters.password = job.authPassword
+            }
+        }
+
+        String apiKey = job.location.wptServer.apiKey
+        if (apiKey) {
+            parameters.k = apiKey
+        }
+
+        parameters.f = 'xml'
+
+        // convert all boolean parameters to 0 or 1
+        parameters = parameters.each {
+            if (it.value instanceof Boolean && it.value != null)
+                it.value = it.value ? 1 : 0
+        }
+        return parameters
+    }
+
+
+    private NodeChild parseResponse(def runtestResponse, WebPageTestServer wptserver) {
+        NodeChild runtestResponseXml
+        if (runtestResponse instanceof NodeChild || runtestResponse instanceof GPathResult) {
+            runtestResponseXml = runtestResponse
+            return runtestResponseXml
+        } else {
+            throw new JobExecutionException("Response is not XML from wptserver ${wptserver}", WptStatus.UNKNOWN, null)
+        }
+    }
 }
